Update ui.py
added user_list and save_users to facilitate user profiles logging in. These are read from and written to (respectively) a json file on the server side that stores user names and their hashed passwords.
added logout function.
updated all functions to default to login page if user not found in session.